sandbox/local_friend/requirements.txt
bleak
pyobjc
pyaudio
